fix: resolve logical duplication after rebase — deduplicate housekeeping functions
After rebasing onto main, housekeeping.py had old versions of functions
while scheduler.py had newer versions (sweep refactor, ghost lease fix,
renew_active_leases). This commit:

- Removes 926 lines of duplicated housekeeping functions from scheduler.py
- Updates housekeeping.py with the latest function versions from main
- Adds renew_active_leases, _task_past_grace, _sweep_task_resources to housekeeping.py
- Fixes ghost lease conditional in check_and_requeue_expired_leases
- Updates test patch targets from octopoid.scheduler to octopoid.housekeeping
  where the code under test now lives in housekeeping.py
- Re-exports find_pid_for_task from scheduler for backward compat

All 1161 unit tests pass.
