+++ tests/test_scheduler_refactor.py
@@ -1455,7 +1455,7 @@ class TestSweepTaskResources:
         mock_result.returncode = 0
         mock_result.stderr = ""
 
-        with patch("octopoid.scheduler.run_git", return_value=mock_result) as mock_git:
+        with patch("octopoid.housekeeping.run_git", return_value=mock_result) as mock_git:
             result = _sweep_task_resources(task, tasks_dir, logs_dir, parent_repo)
 
         assert result is True
@@ -1481,7 +1481,7 @@ class TestSweepTaskResources:
         mock_result.returncode = 0
         mock_result.stderr = ""
 
-        with patch("octopoid.scheduler.run_git", return_value=mock_result):
+        with patch("octopoid.housekeeping.run_git", return_value=mock_result):
             result = _sweep_task_resources(task, tasks_dir, logs_dir, parent_repo)
 
         assert result is False
@@ -1500,7 +1500,7 @@ class TestSweepTaskResources:
         mock_result.returncode = 0
         mock_result.stderr = ""
 
-        with patch("octopoid.scheduler.run_git", return_value=mock_result) as mock_git:
+        with patch("octopoid.housekeeping.run_git", return_value=mock_result) as mock_git:
             _sweep_task_resources(task, tasks_dir, logs_dir, parent_repo)
 
         branch_delete_call = mock_call(
@@ -1523,7 +1523,7 @@ class TestSweepTaskResources:
         mock_result.returncode = 0
         mock_result.stderr = ""
 
-        with patch("octopoid.scheduler.run_git", return_value=mock_result) as mock_git:
+        with patch("octopoid.housekeeping.run_git", return_value=mock_result) as mock_git:
             _sweep_task_resources(task, tasks_dir, logs_dir, parent_repo)
 
         # No git push --delete call
